+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 - Fixed: unable to search by tags
 
+## v0.1.9
+
+- Fixed: Parsing `DateRange` error
+
+- Chore: Use timestamp as `DateRange` unit
+
 ## v0.1.8
 
 - Fixed: Do not import existing blogs repeatedly